Add Chart.js v4 visualizations for 17 financial calculators
- Add Chart.js v4 via CDN with modern doughnut & grouped bar charts
- 16 calculators show doughnut charts: Total Invested vs Final Value
  (SIP, LUMPSUM, STEP_UP_SIP, PPF, NSC, FD, RD, Simple Interest,
   Inflation, EMI×5, Compound Interest, GST)
- FLAT_VS_REDUCING shows grouped bar: Flat vs Reducing EMI comparison
- 8 calculators intentionally without charts (SWP, CAGR, Retirement,
  Gratuity, Salary, Brokerage, EPF, NPS)

- Theme-aware colors: adapts to 5 themes (indigo/green/orange/purple/teal)
  + light/dark mode with CSS variable integration
- Auto-re-render on theme/color-scheme change via wrapped setTheme/setColorScheme
- Dark mode tooltips use --bg-tertiary; light mode uses high-contrast text
- Percentage labels on doughnut segments (e.g., "Total Invested (35%)")
- Modern styling: gradient fills, rounded corners, hover points, smooth 600ms animations

- Fix calculator.py: int() conversion for STEP_UP_SIP/RD loop ranges

+++ calculator.py
@@ -43,7 +43,7 @@ def SWP(total_investment, withdrawal_amount, Expected_rate, years):
 def STEP_UP_SIP(monthly_investment, step_up_rate, Expected_return, years):
     monthly_rate = Expected_return / 12 / 100
     step_up = step_up_rate / 100
-    total_months = years * 12
+    total_months = int(years * 12)
 
     fund_value = 0
     total_investment = 0
@@ -133,7 +133,7 @@ def FD_SIMPLE(principal, interest_rate, years):
 
 # RECURRING DEPOSIT
 def RD(monthly_investment, Expected_rate, years):
-    months = years * 12
+    months = int(years * 12)
     quarterly_rate = Expected_rate / 100 / 4
     maturity = 0
     values = []
